requirements.txt
Django==2.2
django-filter==21.1
djangorestframework==3.12.0
psycopg2-binary>=2.8
python-dotenv==0.20.0
pytz==2022.1
sqlparse==0.4.2
